requirements.txt
pandas>=1.5.0
pyyaml>=6.0.0
numpy<2.0.0  # Using older version for compatibility
openpyxl>=3.0.0  # Required for reading Excel files 